+++ docs/changelog.md
@@ -6,6 +6,7 @@
 
 * fix `validate_default_values` - it now validates with in context of defaults
 * new func `utils.remove_same_ident` and apply it for SimpleYAML
+* fix `SimpleBool` validators
 
 ### 0.28 precached Py values and to_raw